Change the way c_str_vals are managed
Originally, I thought we could handle strings just as any other constants.
Since I started handling them as global values, though, it has become apparent
that it should be possible to have different c_str_vals with the same content.
For once, global values have a parent pointer. If two modules are used with
the same ir_context, it would be impossible to have the same string in both
modules.

This commit also cleans up the ir codegen for '+' so it makes use of some
recently added helper functions