Store all ast nodes in a single deque
Fix #2
On my system, every ast node now takes up 120 bytes (or 15 words). The
largest kind of ast node should be class_expr_node, with 2 vectors (~3 words
each), one optional str_table_entry (~2+1 words) and an optional pointer (~2
words, 'constructor'). Then there is the tag for the variant, which may
need another word, as well as the pointer to the ast_node vtable.
ast_node also has a source_loc member, so that's another 2 words, which
now brings us to 15 words total.

Maybe we can treat str_table_entry itself as an optional, which would save us 1
word. Furthermore, we could turn the vectors into sub-nodes, reducing their
size from 3 to one word each. We also have the possibility to remove either the
variant tag or the ast_node vptr in favor of the other - I think falling
back to a plain union with a visitor-based destructor impl would be the
way to go here. These measures could decrease the size of one ast node
by 6 words (40%).
I wouldn't want to abuse nullptr to get rid of the optional<pointer>,
though, even if it could save another word. Treating nullptr as an error
in the ast is the safer solution in my opinion.