Move ir printing logic into dedicated printer class
In the near future, we want to be able to parse printed IR again, so it
is important to gather all printing logic in one place as to make
reasoning about the text representation easier.
I know that with the new printing implementation, printing of a function
is quadratic in the number of instructions it contains. However, this
can easily be optimized in the future - I just don't want to waste too
much time on it right now.

This commit also adds a boolean to functions to clearly indicate whether
it is a normal function or an intrinsic.
Furthermore, several new read-only apis on ir values have been added to
allow printing without having to make the printer a friend class.